% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{filter_by_expression_rank}
\alias{filter_by_expression_rank}
\title{Filter top N TF scores based on expression rank}
\usage{
filter_by_expression_rank(rep.matrix, tf.scores, query.name, pn = 0.05,
  pnE = 0.05)
}
\arguments{
\item{rep.matrix}{is a expression matrix of the representative samples from prepare_representaive_samples}

\item{tf.scores}{a numeric data.frame containing CaCTS scores.}

\item{query.name}{name of query cancer type or subtype}

\item{pn}{percentage of top N scores to be considered. Default 0.05}

\item{pnE}{percentage of top N expression rank to be considered. Default 0.05}
}
\value{
Return the filtered list of CaCTS scores considering expression rank.
}
\description{
This is .
}
