% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_representaive_samples}
\alias{prepare_representaive_samples}
\title{Prepare the expression matrix to contain representative samples}
\usage{
prepare_representaive_samples(expr.matrix, sample.descr, save.file = T,
  rep.NA = 1e-17,
  filename = "expr.matrix.representative.sample.mean.txt")
}
\arguments{
\item{expr.matrix}{a gene expression matrix}

\item{sample.descr}{a sample annotation data.frame containing cancer type or cancer subtype information}

\item{save.file}{if True the matrix with representative samples with be saved.}

\item{rep.NA}{value to replace NAs. Default 1e-17}

\item{filename}{Name for the outuput expression matrix file.}
}
\value{
Return a expression matrix of the representative samples.
}
\description{
Representative samples are generated considering the expression mean across all samples
}
