% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{visualize_scores}
\alias{visualize_scores}
\title{Vizualize the overall CaCTS scores curve and the expression values of top N ranked CaCTS scores}
\usage{
visualize_scores(tf.scores, rep.matrix, topn = 5, query.name, filename,
  h, w)
}
\arguments{
\item{tf.scores}{a numeric data.frame containing CaCTS scores.}

\item{rep.matrix}{a expression matrix of the representative samples}

\item{topn}{number of top scores TF to be shown (default = 5)}

\item{query.name}{name of query cancer type or subtype}

\item{filename}{Name for the plot outuput saved as pdf}

\item{h}{height value for plot}

\item{w}{width value for plot}
}
\value{
a prot_grid object containing a list of ggplot2 plots
}
\description{
the visualize_scores function helps an user to identify the expression profile of a given TF
across all cancer type.
}
